Moving getContents Back
