+++ src/Views/Columns/ArrayColumn.php
@@ -6,6 +6,9 @@ use Rappasoft\LaravelLivewireTables\Views\Column;
 use Rappasoft\LaravelLivewireTables\Views\Columns\Traits\Configuration\ArrayColumnConfiguration;
 use Rappasoft\LaravelLivewireTables\Views\Columns\Traits\Helpers\ArrayColumnHelpers;
 use Rappasoft\LaravelLivewireTables\Views\Columns\Traits\IsColumn;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\HtmlString;
+use Rappasoft\LaravelLivewireTables\Exceptions\DataTableConfigurationException;
 
 class ArrayColumn extends Column
 {
@@ -31,5 +34,32 @@ class ArrayColumn extends Column
         if (! isset($from)) {
             $this->label(fn () => null);
         }
+    }
+
+    
+    public function getContents(Model $row): null|string|\BackedEnum|HtmlString|DataTableConfigurationException|\Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+    {
+        $outputValues = [];
+        $value = $this->getValue($row);
+
+        if (! $this->hasDataCallback()) {
+            throw new DataTableConfigurationException('You must set a data() method on an ArrayColumn');
+        }
+
+        if (! $this->hasOutputFormatCallback()) {
+            throw new DataTableConfigurationException('You must set an outputFormat() method on an ArrayColumn');
+        }
+
+        foreach (call_user_func($this->getDataCallback(), $value, $row) as $i => $v) {
+            $outputValues[] = call_user_func($this->getOutputFormatCallback(), $i, $v);
+        }
+
+        $returnedValue = (! empty($outputValues) ? implode($this->getSeparator(), $outputValues) : $this->getEmptyValue());
+
+        if ($this->hasOutputWrapperStart() && $this->hasOutputWrapperEnd()) {
+            $returnedValue = $this->getOutputWrapperStart().$returnedValue.$this->getOutputWrapperEnd();
+        }
+
+        return new HtmlString($returnedValue);
     }
 }
