+++ src/Views/Columns/Traits/Helpers/ArrayColumnHelpers.php
@@ -2,9 +2,6 @@
 
 namespace Rappasoft\LaravelLivewireTables\Views\Columns\Traits\Helpers;
 
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\HtmlString;
-use Rappasoft\LaravelLivewireTables\Exceptions\DataTableConfigurationException;
 
 trait ArrayColumnHelpers
 {
@@ -63,29 +60,4 @@ trait ArrayColumnHelpers
         return $this->outputWrapperEnd;
     }
 
-    public function getContents(Model $row): null|string|\BackedEnum|HtmlString|DataTableConfigurationException|\Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
-    {
-        $outputValues = [];
-        $value = $this->getValue($row);
-
-        if (! $this->hasDataCallback()) {
-            throw new DataTableConfigurationException('You must set a data() method on an ArrayColumn');
-        }
-
-        if (! $this->hasOutputFormatCallback()) {
-            throw new DataTableConfigurationException('You must set an outputFormat() method on an ArrayColumn');
-        }
-
-        foreach (call_user_func($this->getDataCallback(), $value, $row) as $i => $v) {
-            $outputValues[] = call_user_func($this->getOutputFormatCallback(), $i, $v);
-        }
-
-        $returnedValue = (! empty($outputValues) ? implode($this->getSeparator(), $outputValues) : $this->getEmptyValue());
-
-        if ($this->hasOutputWrapperStart() && $this->hasOutputWrapperEnd()) {
-            $returnedValue = $this->getOutputWrapperStart().$returnedValue.$this->getOutputWrapperEnd();
-        }
-
-        return new HtmlString($returnedValue);
-    }
 }
